+++ Release/Changelog.md
@@ -1,5 +1,6 @@
 ## v2.6.2 (unreleased)
 - Update version number in Info.cs.
+- Add handling for exceptions thrown while preloading bro assets.
 
 ## v2.6.1
 - Add API methods for allowing other mods to see when a bro is being unlocked and register callbacks.